[BE] 상위 branch Merge Request

See merge request s09-final/S09P31C207!258

+++ server/plonitservice/src/main/java/com/plonit/plonitservice/api/crew/service/impl/CrewServiceImpl.java
@@ -3,8 +3,11 @@ package com.plonit.plonitservice.api.crew.service.impl;
 import com.plonit.plonitservice.api.crew.service.CrewService;
 import com.plonit.plonitservice.api.crew.service.dto.ApproveCrewDto;
 import com.plonit.plonitservice.api.crew.service.dto.SaveCrewDto;
+import com.plonit.plonitservice.api.crew.service.dto.SaveCrewNoticeDto;
+import com.plonit.plonitservice.api.crew.service.dto.UpdateCrewNoticeDto;
 import com.plonit.plonitservice.common.AwsS3Uploader;
 import com.plonit.plonitservice.common.exception.CustomException;
+import com.plonit.plonitservice.common.util.RequestUtils;
 import com.plonit.plonitservice.domain.crew.Crew;
 import com.plonit.plonitservice.domain.crew.CrewMember;
 import com.plonit.plonitservice.domain.crew.repository.CrewMemberQueryRepository;
@@ -14,6 +17,9 @@ import com.plonit.plonitservice.domain.crew.repository.CrewRepository;
 import com.plonit.plonitservice.domain.member.Member;
 import com.plonit.plonitservice.domain.member.repository.MemberQueryRepository;
 import com.plonit.plonitservice.domain.member.repository.MemberRepository;
+import com.plonit.plonitservice.domain.region.Dong;
+import com.plonit.plonitservice.domain.region.Gugun;
+import com.plonit.plonitservice.domain.region.repository.RegionQueryRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -36,6 +42,7 @@ public class CrewServiceImpl implements CrewService {
     private final MemberQueryRepository memberQueryRepository;
     private final MemberRepository memberRepository;
     private final AwsS3Uploader awsS3Uploader;
+    private final RegionQueryRepository regionQueryRepository;
 
     @Override
     @Transactional // 크루 생성
@@ -52,7 +59,11 @@ public class CrewServiceImpl implements CrewService {
                 throw new CustomException(INVALID_FIELDS_REQUEST);
             }
         }
-        Crew crew = crewRepository.save(SaveCrewDto.toEntity(saveCrewDTO, crewImageUrl));
+
+        Dong dong = regionQueryRepository.findDongFetchJoin(saveCrewDTO.getDongCode())
+                .orElseThrow(() -> new CustomException(REGION_NOT_FOUND));
+
+        Crew crew = crewRepository.save(SaveCrewDto.toEntity(saveCrewDTO, crewImageUrl, dong));
         CrewMember crewMember = CrewMember.toEntity(crew, member);
 
         crewMemberRepository.save(crewMember);
@@ -64,15 +75,19 @@ public class CrewServiceImpl implements CrewService {
     public void joinCrewMember(Long memberId, Long crewId) {
         log.info(logCurrent(getClassName(), getMethodName(), START));
 
+        // 멤버 조회
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new CustomException(USER_NOT_FOUND));
 
+        // 크루 조회
         Crew crew = crewRepository.findById(crewId)
                 .orElseThrow(() -> new CustomException(CREW_NOT_FOUND));
 
+        // 크루원 조회
         if (crewMemberQueryRepository.isValidCrewMember(memberId, crewId, false))
             throw new CustomException(CREW_ALREADY_JOIN);
 
+        // 크루원 저장
         CrewMember crewMember = CrewMember.memberToEntity(crew, member);
 
         crewMemberRepository.save(crewMember);
@@ -88,13 +103,98 @@ public class CrewServiceImpl implements CrewService {
         if (!crewMemberQueryRepository.isValidCrewMember(approveCrewDto.getMemberKey(), approveCrewDto.getCrewId(), true))
             throw new CustomException(CREW_MASTER_NOT_FORBIDDEN);
 
-        // 크루원 승인이 대기 중인지 확인
+        // 크루원 확인
         CrewMember crewMember = crewMemberQueryRepository.findByMemberIdAndCrewId(approveCrewDto.getCrewMemberId(), approveCrewDto.getCrewId())
-                .orElseThrow(() -> new CustomException(CREW_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(CREW_MEMBER_NOT_FOUND));
+
+        // 승인 대기 확인
+        if(crewMember.getIsCrewMember())
+            throw new CustomException(CREW_ALREADY_JOIN);
 
         // 승인 -> change , 거절 -> delete
         if (approveCrewDto.getStatus()) crewMember.changeIsCrewMember();
         else crewMemberRepository.delete(crewMember);
+
+        // todo : 크루 승인, 거절 알림
+
+        log.info(logCurrent(getClassName(), getMethodName(), END));
+    }
+
+    @Override
+    @Transactional // 공지사항 등록
+    public void saveCrewNotice(SaveCrewNoticeDto saveCrewNoticeDto) {
+        log.info(logCurrent(getClassName(), getMethodName(), START));
+
+        Long memberId = RequestUtils.getMemberId();
+
+        // 크루장인지 확인
+        if (!crewMemberQueryRepository.isValidCrewMember(memberId, saveCrewNoticeDto.getCrewId(), true))
+            throw new CustomException(CREW_MASTER_NOT_FORBIDDEN);
+
+
+        Crew crew = crewRepository.findById(saveCrewNoticeDto.getCrewId())
+                .orElseThrow(() -> new CustomException(CREW_NOT_FOUND));
+
+        // 공지사항 등록
+        crew.changeNotice(saveCrewNoticeDto.getContent());
+
+        log.info(logCurrent(getClassName(), getMethodName(), END));
+    }
+
+    @Override
+    @Transactional // 크루 탈퇴
+    public void quitCrew(Long crewId) {
+        log.info(logCurrent(getClassName(), getMethodName(), START));
+
+        Long memberId = RequestUtils.getMemberId();
+
+        // 크루 확인
+        if(!crewQueryRepository.isValidCrew(crewId))
+            throw new CustomException(CREW_NOT_FOUND);
+
+        // 크루장인지 확인 -> 탈퇴 X
+        if (crewMemberQueryRepository.isValidCrewMember(memberId, crewId, true))
+            throw new CustomException(CREW_MASTER_NOT_QUIT);
+
+        // 크루원 확인
+        CrewMember crewMember = crewMemberQueryRepository.findByMemberIdAndCrewId(memberId, crewId)
+                .orElseThrow(() -> new CustomException(CREW_MEMBER_NOT_FOUND));
+
+        // 승인 대기 확인
+        if(!crewMember.getIsCrewMember())
+            throw new CustomException(CREW_WAITING_FORBIDDEN);
+
+        crewMemberRepository.delete(crewMember);
+
+        log.info(logCurrent(getClassName(), getMethodName(), END));
+    }
+
+    @Override
+    @Transactional // 크루 강퇴
+    public void kickOutCrew(Long crewMemberId, Long crewId) {
+        log.info(logCurrent(getClassName(), getMethodName(), START));
+
+        Long memberId = RequestUtils.getMemberId();
+
+        // 크루 확인
+        if(!crewQueryRepository.isValidCrew(crewId))
+            throw new CustomException(CREW_NOT_FOUND);
+
+        // (로그인한 유저) 크루장인지 확인
+        if (!crewMemberQueryRepository.isValidCrewMember(memberId, crewId, true))
+            throw new CustomException(CREW_MASTER_NOT_FORBIDDEN);
+
+        // 크루원 확인
+        CrewMember crewMember = crewMemberQueryRepository.findByMemberIdAndCrewId(crewMemberId, crewId)
+                .orElseThrow(() -> new CustomException(CREW_MEMBER_NOT_FOUND));
+
+        // 승인 대기 확인
+        if(!crewMember.getIsCrewMember())
+            throw new CustomException(CREW_WAITING_FORBIDDEN);
+
+        crewMemberRepository.delete(crewMember);
+
+        // todo : 강퇴 알림
 
         log.info(logCurrent(getClassName(), getMethodName(), END));
     }
